requirements.txt
streamlit
openai
google-generativeai
anthropic
langchain
chromadb
pysqlite3-binary